backend/pyproject.toml
[tool.poetry]
name = "inviterr"
version = "0.0.1"
description = ""
authors = ["WardPearce <wardpearce@pm.me>"]
readme = "README.md"
packages = [
    {include = "*", from = "app"},
    {include = "run.py"}
]

[tool.poetry.scripts]
server = "run:main"

[tool.poetry.dependencies]
python = "^3.11"
litestar = {extras = ["full"], version = "^2.14.0"}
aiohttp = "^3.11.11"
motor = "^3.6.1"
bcrypt = "^4.2.1"
plexapi = "^4.16.1"
pydantic-settings = "^2.7.1"
apscheduler = {version = "3.11.0", extras = ["mongodb"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
